EDIT: 사용자 발급 검증 로직과 비지니스 로직 분리(#38)
