+++ src/main/java/com/ayucoupon/usercoupon/service/issue/IssueUserCouponModule.java
@@ -1,0 +1,6 @@
+package com.ayucoupon.usercoupon.service.issue;
+
+public interface IssueUserCouponModule {
+     Long issue(IssueUserCouponCommand command);
+
+}
