+++ src/test/java/com/ayucoupon/usercoupon/controller/UserCouponControllerTest.java
@@ -8,8 +8,8 @@ import com.ayucoupon.usercoupon.controller.request.UseUserCouponRequest;
 import com.ayucoupon.usercoupon.controller.request.IssueUserCouponRequest;
 import com.ayucoupon.usercoupon.domain.entity.UserCoupon;
 import com.ayucoupon.usercoupon.service.*;
-import com.ayucoupon.usercoupon.service.IssueUserCouponCommand;
-import com.ayucoupon.usercoupon.service.IssueUserCouponService;
+import com.ayucoupon.usercoupon.service.issue.IssueUserCouponCommand;
+import com.ayucoupon.usercoupon.service.issue.IssueUserCouponService;
 import com.ayucoupon.util.Coupons;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.BeforeEach;
