+++ src/test/java/com/ayucoupon/usercoupon/service/ShowUserCouponServiceTest.java
@@ -1,6 +1,8 @@
 package com.ayucoupon.usercoupon.service;
 
 
+import com.ayucoupon.usercoupon.service.issue.IssueUserCouponCommand;
+import com.ayucoupon.usercoupon.service.issue.IssueUserCouponService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
